neon, app: migrate branding into the theme
