+++ packages/neon/neon/lib/models.dart
@@ -1,4 +1,3 @@
 export 'package:neon/src/models/account.dart';
 export 'package:neon/src/models/app_implementation.dart';
-export 'package:neon/src/models/branding.dart';
 export 'package:neon/src/models/notifications_interface.dart';
