+++ packages/neon/neon/lib/neon.dart
@@ -10,8 +10,8 @@ import 'package:neon/src/blocs/next_push.dart';
 import 'package:neon/src/blocs/push_notifications.dart';
 import 'package:neon/src/models/account.dart';
 import 'package:neon/src/models/app_implementation.dart';
-import 'package:neon/src/models/branding.dart';
 import 'package:neon/src/platform/platform.dart';
+import 'package:neon/src/theme/neon.dart';
 import 'package:neon/src/utils/global_options.dart';
 import 'package:neon/src/utils/request_manager.dart';
 import 'package:package_info_plus/package_info_plus.dart';
@@ -24,7 +24,7 @@ late final String neonUserAgent;
 Future runNeon({
   required final Iterable<AppImplementation> Function(SharedPreferences, RequestManager, NeonPlatform)
       getAppImplementations,
-  required final Branding branding,
+  required final NeonTheme theme,
   final WidgetsBinding? bindingOverride,
   final SharedPreferences? sharedPreferencesOverride,
   final Account? account,
@@ -115,11 +115,8 @@ Future runNeon({
         Provider<PackageInfo>(
           create: (final _) => packageInfo,
         ),
-        Provider<Branding>(
-          create: (final _) => branding,
-        ),
       ],
-      child: const NeonApp(),
+      child: NeonApp(neonTheme: theme),
     ),
   );
 }
