+++ packages/neon/neon/lib/src/pages/settings.dart
@@ -8,7 +8,6 @@ import 'package:neon/l10n/localizations.dart';
 import 'package:neon/src/blocs/accounts.dart';
 import 'package:neon/src/models/account.dart';
 import 'package:neon/src/models/app_implementation.dart';
-import 'package:neon/src/models/branding.dart';
 import 'package:neon/src/platform/platform.dart';
 import 'package:neon/src/router.dart';
 import 'package:neon/src/settings/widgets/account_settings_tile.dart';
@@ -19,6 +18,7 @@ import 'package:neon/src/settings/widgets/settings_category.dart';
 import 'package:neon/src/settings/widgets/settings_list.dart';
 import 'package:neon/src/settings/widgets/settings_tile.dart';
 import 'package:neon/src/settings/widgets/text_settings_tile.dart';
+import 'package:neon/src/theme/branding.dart';
 import 'package:neon/src/utils/confirmation_dialog.dart';
 import 'package:neon/src/utils/global_options.dart';
 import 'package:neon/src/utils/save_file.dart';
@@ -228,7 +228,7 @@ class _SettingsPageState extends State<SettingsPage> {
                         ),
                         title: Text(AppLocalizations.of(context).licenses),
                         onTap: () async {
-                          final branding = Provider.of<Branding>(context, listen: false);
+                          final branding = Branding.of(context);
                           showLicensePage(
                             context: context,
                             applicationName: branding.name,
