+++ packages/neon/neon/lib/theme.dart
@@ -1,0 +1,4 @@
+export 'src/theme/branding.dart';
+export 'src/theme/color_scheme.dart';
+export 'src/theme/dialog.dart';
+export 'src/theme/neon.dart';
